[http] Add optional additional headers and other changes
Support for optionally injecting additional headers into the HTTP request,
which are passed by a caller of http_open_filter().  This can be useful for
requesting byte ranges, for example.

Support for parsing Content-Range headers in responses.  If a caller to
http_open_filter() passes us a pointer to a resource length field, we will
fill this field in on their behalf during processing.  Such a caller can
then know the total length of a resource, even if they only requested a byte
range.