Allow building for opt-out boot strategy
We detect the VMware Player and VMware Workstation 7.0.1
BIOS as a non-BBS-compliant BIOS and we therefore hook
INTerrupt 0x19 for booting.  As of commit
4d7c650164a759e3dadbcf8f83da6789165c68b7, we would choose
to always boot unless the user at the console specifies
otherwise with an 'N' keystroke.  Some users would
rather that we not always boot by default.

This commit introduces a conditional build path for
allowing INT 0x19 users to build an "opt-out" ROM.
If INT19_OPTOUT is defined, such a ROM image will result.

This commit also adds a ROM-O-Matic option for the same.

Reported-by: Alessandro Salvatori <sandr8@gmail.com>
Tested-by: Alessandro Salvatori <sandr8@gmail.com>
Tested-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>
Signed-off-by: Shao Miller <shao.miller@yrdsb.edu.on.ca>